Completed monitoring mode selection both Single and 4 Cameras
